Separate package and dispatcher for PieConfig
