+++ shortcut_composer/templates/pie_menu_utils/__init__.py
@@ -3,7 +3,7 @@
 
 """Implementation of PieMenu main elements."""
 
-from .pie_config import PieConfig, PresetPieConfig, NonPresetPieConfig
+from .pie_config import PieConfig
 from .pie_settings import PieSettings
 from .label_widget import LabelWidget
 from .pie_manager import PieManager
@@ -14,8 +14,6 @@ from .edit_mode import EditMode
 from .label import Label
 
 __all__ = [
-    "NonPresetPieConfig",
-    "PresetPieConfig",
     "PieSettings",
     "LabelWidget",
     "PieConfig",
