+++ shortcut_composer/templates/pie_menu_utils/pie_settings_impl/enum_group_pie_settings.py
@@ -6,11 +6,8 @@ from enum import Enum
 
 from api_krita.enums.helpers import EnumGroup
 from core_components import Controller
-from templates.pie_menu_utils import (
-    Label,
-    PieStyle,
-    PieSettings,
-    NonPresetPieConfig)
+from templates.pie_menu_utils.pie_config_impl import NonPresetPieConfig
+from templates.pie_menu_utils import Label, PieStyle, PieSettings
 from .common_utils import GroupScrollArea, GroupManager
